Fix TypeScript errors and add documentation
- Fixed server initialization method (start -> connect)
- Corrected import paths for RequestHandlerExtra
- Fixed McpToolResponse interface
- Added type assertions to string literals
- Updated .eslintrc.json for unused parameters
- Added README.md documentation for directory structure
- Fixed linting issues across source files
