+++ .eslintrc.json
@@ -13,6 +13,7 @@
   "rules": {
     "no-console": ["warn", { "allow": ["error", "warn"] }],
     "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/no-explicit-any": "off"
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-unused-vars": ["error", { "argsIgnorePattern": "^_" }]
   }
 }
