+++ package.json
@@ -31,6 +31,6 @@
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "prettier": "^3.0.0",
-    "typescript": "^5.0.0"
+    "typescript": "~5.3.3"
   }
 }
