+++ src/tools/branches/README.md
@@ -1,0 +1,16 @@
+# Branches Tools
+
+This directory is reserved for future implementation of tools for managing Xano branches.
+
+## Planned Features
+
+- List branches
+- Create branch
+- Update branch
+- Delete branch
+- Merge branch
+- Set default branch
+
+## Implementation Status
+
+These tools are not yet implemented. Check back in future versions.
