+++ src/tools/columns/README.md
@@ -1,0 +1,15 @@
+# Columns Tools
+
+This directory is reserved for future implementation of tools for managing Xano table columns.
+
+## Planned Features
+
+- Add column to table
+- Remove column from table
+- Modify column properties
+- List columns in a table
+- Get column details
+
+## Implementation Status
+
+These tools are not yet implemented. Column operations are currently handled through the table schema tools in the tables directory. Check back in future versions for dedicated column tools.
