+++ src/tools/functions/README.md
@@ -1,0 +1,16 @@
+# Functions Tools
+
+This directory is reserved for future implementation of tools for managing Xano functions.
+
+## Planned Features
+
+- List functions
+- Get function details
+- Create function
+- Update function
+- Delete function
+- Test function execution
+
+## Implementation Status
+
+These tools are not yet implemented. Check back in future versions.
