+++ src/tools/tasks/README.md
@@ -1,0 +1,17 @@
+# Tasks Tools
+
+This directory is reserved for future implementation of tools for managing Xano tasks.
+
+## Planned Features
+
+- List tasks
+- Get task details
+- Create task
+- Update task
+- Delete task
+- Execute task
+- Get task execution history
+
+## Implementation Status
+
+These tools are not yet implemented. Check back in future versions.
