+++ src/tools/utils/README.md
@@ -1,0 +1,15 @@
+# Utils
+
+This directory is reserved for future utility functions and helper modules.
+
+## Planned Features
+
+- Additional data transformation utilities
+- Schema validation utilities
+- API response formatters
+- Error handling utilities
+- Pagination helpers
+
+## Implementation Status
+
+These utilities are not yet implemented. Current utility functions are located in the parent directory's utils.ts file. This directory will be used for more complex utilities that warrant their own modules.
