+++ src/tools/workspaces/README.md
@@ -1,0 +1,16 @@
+# Workspaces Tools
+
+This directory is reserved for future implementation of tools for managing Xano workspaces.
+
+## Planned Features
+
+- List workspaces
+- Get workspace details
+- Create workspace
+- Update workspace
+- Delete workspace
+- Clone workspace
+
+## Implementation Status
+
+These tools are not yet implemented. The current implementation uses a single workspace specified by the XANO_WORKSPACE environment variable. Check back in future versions for multi-workspace support.
